+++ .eslintrc.js
@@ -17,7 +17,7 @@ module.exports = {
     sourceType: 'module',
     project: ['./tsconfig.eslint.json'],
   },
-  plugins: ['@typescript-eslint', 'prettier', 'jest'],
+  plugins: ['@typescript-eslint', 'prettier', 'jest', 'jest-formatting'],
   rules: {
     /**********/
     /** Style */
@@ -96,6 +96,10 @@ module.exports = {
 
     // We allow import cycles because they cause no issues for types.
     'import/no-cycle': 'off',
+
+    // padding around describe and tests
+    "jest-formatting/padding-around-describe-blocks": 2,
+    "jest-formatting/padding-around-test-blocks": 2
   },
   overrides: [
     ...readdirSync(resolve(__dirname, 'packages'))