+++ packages/datasource-sql/test/_helpers/setup-using-all-types.ts
@@ -23,7 +23,7 @@ export function getAttributeMapping(dialect: Dialect) {
   return {
     primitiveT: {
       id: {
-        type: DataTypes.NUMBER(),
+        type: DataTypes.INTEGER(),
         allowNull: false,
         primaryKey: true,
         autoIncrement: true,
@@ -39,7 +39,7 @@ export function getAttributeMapping(dialect: Dialect) {
         defaultValue: true,
       },
       int: {
-        type: DataTypes.NUMBER(),
+        type: DataTypes.INTEGER(),
         allowNull: true,
         defaultValue: 2,
       },
@@ -74,7 +74,7 @@ export function getAttributeMapping(dialect: Dialect) {
     },
     paranoidT: {
       id: {
-        type: DataTypes.NUMBER(),
+        type: DataTypes.INTEGER(),
         allowNull: false,
         primaryKey: true,
         autoIncrement: true,
@@ -96,7 +96,7 @@ export function getAttributeMapping(dialect: Dialect) {
     },
     enumT: {
       id: {
-        type: DataTypes.NUMBER(),
+        type: DataTypes.INTEGER(),
         allowNull: false,
         primaryKey: true,
         autoIncrement: true,
@@ -109,7 +109,7 @@ export function getAttributeMapping(dialect: Dialect) {
     },
     jsonT: {
       id: {
-        type: DataTypes.NUMBER(),
+        type: DataTypes.INTEGER(),
         allowNull: false,
         primaryKey: true,
         autoIncrement: true,
@@ -122,7 +122,7 @@ export function getAttributeMapping(dialect: Dialect) {
     },
     jsonBT: {
       id: {
-        type: DataTypes.NUMBER(),
+        type: DataTypes.INTEGER(),
         allowNull: false,
         primaryKey: true,
         autoIncrement: true,
@@ -135,13 +135,13 @@ export function getAttributeMapping(dialect: Dialect) {
     },
     arrayT: {
       id: {
-        type: DataTypes.NUMBER(),
+        type: DataTypes.INTEGER(),
         allowNull: false,
         primaryKey: true,
         autoIncrement: true,
       },
       arrayInt: {
-        type: DataTypes.ARRAY(DataTypes.NUMBER),
+        type: DataTypes.ARRAY(DataTypes.INTEGER),
         allowNull: true,
       },
       arrayString: {