+++ packages/datasource-toolkit/src/interfaces/schema.ts
@@ -34,6 +34,7 @@ export type ColumnSchema = {
   isPrimaryKey?: boolean;
   isReadOnly?: boolean;
   isSortable?: boolean;
+  allowNull?: boolean;
   type: 'Column';
   validation?: ColumnSchemaValidation;
 };