+++ packages/datasource-mongoose/CHANGELOG.md
@@ -1,3 +1,33 @@
+## @forestadmin/datasource-mongoose [1.6.3](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-mongoose@1.6.2...@forestadmin/datasource-mongoose@1.6.3) (2024-02-06)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.32.0
+
+## @forestadmin/datasource-mongoose [1.6.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-mongoose@1.6.1...@forestadmin/datasource-mongoose@1.6.2) (2024-02-02)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.31.0
+
+## @forestadmin/datasource-mongoose [1.6.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-mongoose@1.6.0...@forestadmin/datasource-mongoose@1.6.1) (2024-02-01)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.30.1
+
 # @forestadmin/datasource-mongoose [1.6.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-mongoose@1.5.33...@forestadmin/datasource-mongoose@1.6.0) (2024-01-26)